Extracted SubsumptionTableEntry::ApplySubstitutionVisitor to TxSubstitutionVisitor in include/klee/util/TxExprUtil.h
